+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 jsu updated to a new version (see documentation): all changes to the library
 must be taken into account when accessing its APIs explicitly, i.e. not via nvc.
 
-- Add `Nvc.parseFsmCsv(str, allowFullBlanks, allowDuplicates)`.
+- Add `Nvc.fsm.parseFsmCsv(str, allowFullBlanks, allowDuplicates)` for parsing
+CSV in FSM context.
+- Add `Nvc.quick.revertOutputFocusListeners()` to revert `Nvc.quick.installOutputFocusListeners()`.
 - Remove `Nvc.convertLatexShortcuts(text)` which should be replaced with `Jsu.Latex.convertLatexShortcuts(text)`.
 - Remove `Nvc.getLatexShortcutPatterns()` which should be replaced with `Jsu.Latex.getLatexShortcutData()`,
 keeping in mind the following changes for the properties of the returned object:
@@ -14,7 +16,7 @@ keeping in mind the following changes for the properties of the returned object:
 `greekLetter.value`       &rarr; `greekLetter.pattern.value`,
 `subscript.specialChar`   &rarr; `subscript.pattern.specialChar`,
 `subscript.value`         &rarr; `subscript.pattern.value`.
-- Update `Nvc.buildFsmTransitionTableHtml(model, htmlAttrs, indents)`: the
+- Update `Nvc.fsm.buildFsmTransitionTableHtml(model, htmlAttrs, indents)`: the
 attributes provided in `htmlAttrs` will always be trimmed before use.
 
 ## 0.9.0 - 2022/09/03